nodemon + super_admin profile

+++ routes/user.routes.js
@@ -69,16 +69,34 @@ router.put('/users/:id', authGuardMiddleware, upload.single('buffer'), async (re
 
 // Delete user by ID - Admin access only!
 router.delete('/users/:id', authGuardMiddleware, async (req, res) => {
-    console.log('DELETE User begin', req.params)
     try {
-        const deletedUser = await User.findByIdAndRemove(req.params.id)
-        console.log('DELETE USER', deletedUser)
-        if (!deletedUser) {
+        // RULE - No self-delete
+        if (req.userId === req.params.id) {
+            console.log('No delete!')
+            return res.status(403).json({ message: "A user can't self-delete." })
+        }
+        const deleteEntity = await User.findById(req.userId)
+        const deleteTarget = await User.findById(req.params.id)
+        console.log('deleteEntity -', deleteEntity)
+        console.log('deleteTarget -', deleteTarget)
+        let deletedUser
+        if (!deleteTarget) {
             return res.status(404).json({ message: 'User not found' })
+        }
+        if (deleteTarget.profile==='super_admin') {
+            // RULE - No superAdmin delete
+            console.log('SuperAdmin - No delete!')
+            return res.status(403).json({ message: "SuperAdmin can't be deleted, he is God!" })
+        } else if (deleteTarget.profile==='admin' && deleteEntity.profile!=='super_admin') {
+            // RULE - No admin delete from other than superAdmin
+            console.log('Admin protected - No delete!')
+            return res.status(403).json({ message: "Only SuperAdmin can delete an Admin!" })
+        } else {
+            deletedUser = await User.findByIdAndRemove(req.params.id)
         }
         res.json(deletedUser)
     } catch (err) {
-        res.status(500).json({ message: 'Unable to delete user', err })
+        res.status(500).json({ message: 'Unable to delete user'})
     }
 })
